requirements.txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
boto3==1.34.34
botocore==1.34.162
cachetools==5.5.2
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.8.0
fonttools==4.61.1
fpdf2==2.8.5
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
importlib_metadata==7.2.1
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.8.2
mdurl==0.1.2
narwhals==2.16.0
numpy==1.26.4
packaging==23.2
pandas==2.1.4
patsy==1.0.2
pillow==10.4.0
plotly==6.5.2
plotly-express==0.4.1
protobuf==4.25.8
pyarrow==23.0.0
pydeck==0.9.1
Pygments==2.19.2
pymongo==4.6.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
reportlab==4.0.8
requests==2.31.0
rich==13.9.4
rpds-py==0.30.0
s3transfer==0.10.4
scipy==1.17.0
six==1.17.0
smmap==5.0.2
statsmodels==0.14.6
streamlit==1.53.1
tenacity==8.5.0
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
validators==0.35.0
watchdog==6.0.0
zipp==3.23.0